Add files via upload
RAM usage for Johnson-Lindenstrauss lemma to solve for effective resistance has been (mostly) mostly fixed. It should be noted that the dimensionality reduction matrix Z still is large. For very large networks, it could be that this should be written to disk, then read back. For medium-sized to large networks, current methods should still be adequate.